+++ docs/make.jl
@@ -1,11 +1,17 @@
 using Documenter
 using MixedModels
+using FiniteDiff
+using ForwardDiff
 using StatsAPI
 using StatsBase
 
 makedocs(;
     sitename="MixedModels",
+    format=Documenter.HTML(; size_threshold=500_000, size_threshold_warn=250_000),
     doctest=true,
+    # pagesonly=true,
+    # warnonly=true,
+    # warnonly=[:cross_references],
     pages=[
         "index.md",
         "constructors.md",
@@ -15,6 +21,7 @@ makedocs(;
         "bootstrap.md",
         "rankdeficiency.md",
         "mime.md",
+        "derivatives.md",
         "api.md",
     ],
 )