+++ test/pivot.jl
@@ -5,18 +5,20 @@ import MixedModels: statsrank
 xtx(X) = Symmetric(X'X, :U)  # creat the symmetric matrix X'X from X
 # this is defined in Julia 1.13
 @static if VERSION < v"1.13.0-DEV.655"
-    LinearAlgebra.rank(F::LinearAlgebra.QRPivoted; tol=1e-8) = searchsortedlast(abs.(diag(F.R)), tol, rev=true)
+    function LinearAlgebra.rank(F::LinearAlgebra.QRPivoted; tol=1e-8)
+        return searchsortedlast(abs.(diag(F.R)), tol; rev=true)
+    end
 end
 
 const rng = StableRNG(4321234)
 
 const simdat = (
-    G = repeat('A':'T', inner=10),
-    H = repeat('a':'e', inner=2, outer=20),
-    U = repeat(0.:9, outer=20),
-    V = repeat(-4.5:4.5, outer=20),
-    Y = 0.1 * randn(rng, 200),
-    Z = rand(rng, 200)
+    G=repeat('A':'T'; inner=10),
+    H=repeat('a':'e'; inner=2, outer=20),
+    U=repeat(0.0:9; outer=20),
+    V=repeat(-4.5:4.5; outer=20),
+    Y=0.1 * randn(rng, 200),
+    Z=rand(rng, 200),
 )
 
 @testset "fullranknumeric" begin
@@ -26,7 +28,7 @@ const simdat = (
 end
 
 @testset "fullrankcategorical" begin
-    mm = modelmatrix(@formula(Y ~ 1 + G*H), simdat)
+    mm = modelmatrix(@formula(Y ~ 1 + G * H), simdat)
     r, pivot = statsrank(mm)
     @test r == 100
     @test pivot == 1:100
@@ -46,7 +48,7 @@ end
 end
 
 @testset "qr missing cells" begin
-    mm = modelmatrix(@formula(Y ~ 1 + G*H), simdat)[5:end,:]
+    mm = modelmatrix(@formula(Y ~ 1 + G * H), simdat)[5:end, :]
     r, pivot = statsrank(mm)
     @test r == 98
     # we no longer offer ordering guarantees besides preserving